// src/localization/types.ts
export type Language = 'ko' | 'en' | /* 'zh-CN' | 'zh-TW' | */ 'ja' | 'id' | 'es' | 'fr';